self-paced-labs/cloud-hero/optional-kubernetes-engine/requirements.txt
Flask==1.0.4
google-cloud-datastore==1.7.1
google-cloud-storage==1.23.0
google-cloud-logging==1.8.0
google-cloud-error_reporting==0.30.0
gunicorn==19.9.0
oauth2client==4.1.3
mock==2.0.0
Flask-SQLAlchemy==2.3.2
PyMySQL==0.9.2
Flask-PyMongo==2.3.0
PyMongo==3.7.2
six==1.11.0
requests[security]==2.21.0
honcho==1.0.1
psq==0.7.0
